refactor: add response entities PersonController

+++ src/main/java/com/openclassrooms/safetyNet/controller/PersonController.java
@@ -4,6 +4,9 @@ import com.openclassrooms.safetyNet.model.Person;
 import com.openclassrooms.safetyNet.service.PersonService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -15,12 +18,13 @@ public class PersonController {
     private PersonService personService;
 
     @PostMapping("/person")
-    public Person createPerson(@RequestBody Person person) {
-        return personService.savePerson(person);
+    public ResponseEntity<Person> createPerson(@RequestBody Person person) {
+        Person createdPerson = personService.savePerson(person);
+        return ResponseEntity.status(HttpStatus.CREATED).body(createdPerson);
     }
 
     @PutMapping("/person/{lastname}/{firstname}")
-    public Person updatePerson(@PathVariable("lastname") String lastname, @PathVariable("firstname") String firstname, @RequestBody Person person) {
+    public ResponseEntity<?> updatePerson(@PathVariable("lastname") String lastname, @PathVariable("firstname") String firstname, @RequestBody Person person) {
         Optional<Person> p = personService.getPersonByLastnameAndFirstname(lastname, firstname);
         if(p.isPresent()) {
             Person currentPerson = p.get();
@@ -50,27 +54,39 @@ public class PersonController {
                 currentPerson.setEmail(email);
             }
 
-            return personService.savePerson(currentPerson);
+            Person updatedPerson = personService.savePerson(currentPerson);
+            return ResponseEntity.status(HttpStatus.OK).body(updatedPerson);
 
         } else {
-            return null;
+            String notFoundMessage = firstname + " " + lastname + " not found in database";
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(notFoundMessage);
         }
     }
 
     @DeleteMapping("/person")
     @Transactional
-    public void deletePerson(@RequestParam("lastname") String lastname, @RequestParam("firstname") String firstname) {
-        personService.deletePerson(lastname, firstname);
+    public ResponseEntity<?> deletePerson(@RequestParam("lastname") String lastname, @RequestParam("firstname") String firstname) {
+        Optional<Person> p = personService.getPersonByLastnameAndFirstname(lastname, firstname);
+        if(p.isPresent()) {
+            personService.deletePerson(lastname, firstname);
+            String successMessage = firstname + " " + lastname + " has been deleted successfully";
+            return ResponseEntity.status(HttpStatus.OK).body(successMessage);
+        } else {
+            String notFoundMessage = firstname + " " + lastname + " not found in database";
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(notFoundMessage);
+        }
     }
 
     @GetMapping("/communityEmail")
-    public Iterable<String> getCommunityEmail(@RequestParam("city") String city) {
-        return personService.getCommunityEmail(city);
+    public ResponseEntity<?> getCommunityEmail(@RequestParam("city") String city) {
+        Iterable<String> communityEmail = personService.getCommunityEmail(city);
+        return ResponseEntity.status(HttpStatus.OK).body(communityEmail);
     }
 
     @GetMapping("/phoneAlert")
-    public Iterable<String> getPhoneByStationNumber(@RequestParam("firestation") Integer station) {
-        return personService.getPhoneByStationNumber(station);
+    public ResponseEntity<?> getPhoneByStationNumber(@RequestParam("firestation") Integer station) {
+        Iterable<String> phoneNumbers = personService.getPhoneByStationNumber(station);
+        return ResponseEntity.status(HttpStatus.OK).body(phoneNumbers);
     }
 
 }
